+++ src/PS.FarNet.Redis/About/ExportJson.cs
@@ -1,6 +1,8 @@
 ﻿using StackExchange.Redis;
 using System;
 using System.IO;
+using System.Linq;
+using System.Text;
 using System.Text.Encodings.Web;
 using System.Text.Json;
 
@@ -8,33 +10,51 @@ namespace PS.FarNet.Redis;
 
 class ExportJson(string Path, IDatabase Database)
 {
+    public const string KeyBlob = "Blob";
+    public const string KeyText = "Text";
+    public const string KeyHash = "Hash";
+    public const string KeyList = "List";
+    public const string KeySet = "Set";
+    public const string KeyEol = "EOL";
+
     public string Pattern { get; set; }
     public TimeSpan? TimeToLive { get; set; }
     public Action<string> WriteWarning { get; set; }
 
+    static object GetBlobOrText(RedisValue value)
+    {
+        var blob = (byte[])value;
+        var text = Encoding.UTF8.GetString(blob);
+        if (text.IsAscii())
+            return text;
+
+        var temp = Encoding.UTF8.GetBytes(text);
+        return blob.SequenceEqual(temp) ? text : blob;
+    }
+
     static void WriteBlobOrText(Utf8JsonWriter writer, RedisValue value)
     {
-        var text = (string)value;
-        if (text.IsAscii())
+        var obj = GetBlobOrText(value);
+        if (obj is string text)
         {
-            writer.WriteString("Text", text);
+            writer.WriteString(KeyText, text);
         }
         else
         {
-            writer.WriteBase64String("Blob", (byte[])value);
+            writer.WriteBase64String(KeyBlob, (byte[])obj);
         }
     }
 
-    static void WriteString(Utf8JsonWriter writer, RedisValue value, bool isList=false)
+    static void WriteString(Utf8JsonWriter writer, RedisValue value, bool isList = false)
     {
-        var text = (string)value;
-        if (text.IsAscii())
+        var obj = GetBlobOrText(value);
+        if (obj is string text)
         {
             writer.WriteStringValue(text);
         }
         else
         {
-            var base64 = Convert.ToBase64String(value);
+            var base64 = Convert.ToBase64String((byte[])obj);
             var json = isList ? $"{Environment.NewLine}      [\"{base64}\"]" : $"[\"{base64}\"]";
             writer.WriteRawValue(json, true);
         }
@@ -43,7 +63,7 @@ class ExportJson(string Path, IDatabase Database)
     static void WriteEndOfLife(Utf8JsonWriter writer, TimeSpan? ttl)
     {
         if (ttl.HasValue)
-            writer.WriteString("EOL", (DateTime.UtcNow + ttl.Value).ToString("yyyy-MM-dd HH:mm"));
+            writer.WriteString(KeyEol, (DateTime.UtcNow + ttl.Value).ToString("yyyy-MM-dd HH:mm"));
     }
 
     void WriteValue(Utf8JsonWriter writer, RedisKey key, RedisType type, TimeSpan? ttl)
@@ -82,7 +102,7 @@ class ExportJson(string Path, IDatabase Database)
                     writer.WriteStartObject();
                     WriteEndOfLife(writer, ttl);
 
-                    writer.WritePropertyName("Hash");
+                    writer.WritePropertyName(KeyHash);
                     writer.WriteStartObject();
                     foreach (HashEntry entry in res)
                     {
@@ -104,7 +124,7 @@ class ExportJson(string Path, IDatabase Database)
                     writer.WriteStartObject();
                     WriteEndOfLife(writer, ttl);
 
-                    writer.WritePropertyName("List");
+                    writer.WritePropertyName(KeyList);
                     writer.WriteStartArray();
                     foreach (RedisValue item in res)
                         WriteString(writer, item, true);
@@ -123,7 +143,7 @@ class ExportJson(string Path, IDatabase Database)
                     writer.WriteStartObject();
                     WriteEndOfLife(writer, ttl);
 
-                    writer.WritePropertyName("Set");
+                    writer.WritePropertyName(KeySet);
                     writer.WriteStartArray();
                     foreach (RedisValue item in res)
                         WriteString(writer, item, true);